Moving llama training script
